% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PFE.R
\name{rdist}
\alias{rdist}
\title{simulation d'une loi discrete}
\usage{
rdist(x, p)
}
\arguments{
\item{x}{numeric vector represent the values of the random}

\item{p}{numeric vector represent the probabilities}
}
\description{
simulation d'une loi discrete
}
