% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulationCMTD.R
\name{simulationCMTD}
\alias{simulationCMTD}
\title{Simulation Chaine De Markov a temps Discret}
\usage{
simulationCMTD(y, mu, p, n)
}
\arguments{
\item{y}{Vecteur espace des etats}

\item{mu}{Vecteur de distribution initial}

\item{p}{Matrice de transition}

\item{n}{Nombre d'etape}
}
\description{
Simulation Chaine De Markov a temps Discret
}
